pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "moral-path-dependence"
version = "0.1.0"
description = "LLM moral judgment path-dependence benchmark"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "litellm>=1.30",
    "duckdb>=0.10",
    "aiohttp>=3.9",
    "statsmodels>=0.14",
    "scipy>=1.12",
    "numpy>=1.26",
    "pandas>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
analysis = [
    "pymer4>=0.8",   # wraps R lme4; requires R installed separately
    "matplotlib>=3.8",
    "seaborn>=0.13",
]

[project.scripts]
moral-pipeline = "src.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "analysis*"]
